fix: 修复 DPO 训练 loss 不稳定问题
- log_softmax 在 float32 下计算，避免 bfloat16 精度不足导致数值溢出
- 学习率从 5e-5 降至 1e-5，DPO 对参数更新敏感，过大学习率导致 loss 震荡
- global_step 改为仅在梯度累积完成后递增，学习率调度基于实际优化步数
- total_steps 按优化步数计算（mini-steps / accumulation_steps），与调度器同步

+++ local-server/shared/llm/logits_to_log_probs.py
@@ -15,7 +15,8 @@ def logits_to_log_probs(logits, labels):
     Returns:
         每个位置的对数概率, shape [batch, seq_len]
     """
-    log_probs = F.log_softmax(logits, dim=2)
+    # 在 float32 下计算 log_softmax，避免 bfloat16 精度不足导致数值溢出
+    log_probs = F.log_softmax(logits.float(), dim=2)
     log_probs_per_token = torch.gather(log_probs, dim=2, index=labels.unsqueeze(2)).squeeze(-1)
     return log_probs_per_token
